requirements.txt
datasets==3.0.1
huggingface-hub>=0.34.0
pandas
scikit-learn==1.6
streamlit==1.38.0
python-dotenv==1.0.0
matplotlib
seaborn
numpy
joblib
